backend/requirements.txt
Django==5.1.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
django-filter==23.5
